+++ zfit/settings.py
@@ -9,13 +9,44 @@ from .util.execution import RunManager
 run = RunManager()
 
 
-def set_seed(seed):
-    """Set random seed for numpy and TensorFlow."""
-    np.random.seed(seed)
+def set_seed(seed=None, numpy=None, backend=None):
+    """Set random seed for zfit, numpy and the backend.
+
+    Args:
+        seed (int, optional): Seed to set for the random number generator within zfit. If `None` (default), the seed is set to a random value.
+        numpy (int, bool, optional): Seed to set for numpy. If `True` (default), the same seed as for zfit is used. If `False`, the seed is not set.
+        backend (int, bool, optional): Seed to set for the backend. If `True` (default), the same seed as for zfit is used. If `False`, the seed is not set.
+    """
+
     if seed is None:
-        seed = tf.random.get_global_generator().make_seeds(1)
-    tf.random.get_global_generator().reset_from_seed(seed)
-    # tf.random.set_seed(seed)
+        import os
+
+        random_data = os.urandom(4)
+        backend_seed = int.from_bytes(random_data, byteorder="big")
+        backend_seed = int(
+            abs(backend_seed) % 2**31
+        )  # make sure it's positive and not too large
+        seed = backend_seed
+    else:
+        backend_seed = seed
+    if numpy is None:
+        numpy = True
+    if backend is None:
+        backend = True
+
+    if numpy is True:
+        numpy = seed
+    if backend is True:
+        backend = backend_seed
+
+    if numpy is not None:
+        np.random.seed(numpy)
+    if backend is not None:
+        tf.random.set_seed(backend)
+
+    from .z.random import get_prng
+
+    get_prng().reset_from_seed(backend_seed)
 
 
 _verbosity = 0