+++ CHANGELOG.rst
@@ -20,6 +20,9 @@ Bug fixes and small changes
 ---------------------------
 - added ``extended`` as a parameter to all PDFs: a PDF can now directly be extended without the need for
   ``create_extended`` (or ``set_yield``).
+- add ``numpy`` and ``backend`` to options when setting the seed
+- reproducibility by fixing the seed in zfit is restored, ``zfit.run.set_seed`` now also sets the seed for the backend(numpy, tensorflow, etc.) if requested (on by default)
+
 
 Experimental
 ------------
@@ -29,6 +32,7 @@ Requirement changes
 
 Thanks
 ------
+- Sebastian Schmitt @schmitt for reporting the bug in the non-reproducibility of the seed.
 
 0.12.0 (13 March 2023)
 ========================